backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yfinance==0.2.32
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
apscheduler==3.10.4
python-telegram-bot==20.3
anthropic>=0.20.0
pytz==2024.1
